trading_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
yfinance==0.2.33
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
ccxt==4.1.50
asyncio-throttle==1.0.2
websockets==12.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0
prometheus-client==0.19.0
PyJWT==2.8.0
python-binance==1.0.19
psycopg2-binary==2.9.9
gunicorn==21.2.0
aiofiles==23.2.1